+++ .gitignore
@@ -2,3 +2,4 @@
 *.csv
 vendor
 phpunit.xml
+bin
