+++ src/AmazonWishlistExporter/Command/ExportCommand.php
@@ -3,6 +3,7 @@
 namespace AmazonWishlistExporter\Command;
 
 use AmazonWishlistExporter\Logger\LoggerInterface;
+use AmazonWishlistExporter\Crawler\AmazonCrawler;
 use GuzzleHttp\ClientInterface;
 use Symfony\Component\DomCrawler\Crawler;
 
@@ -28,88 +29,45 @@ class ExportCommand implements CommandInterface
      */
     private $logger;
 
+    /**
+     * @var string
+     */
+    private $pathToSave;
+
     public function __construct(
         $countryCode,
         $wishlistId,
         ClientInterface $client,
-        LoggerInterface $logger
-    ) {
+        LoggerInterface $logger,
+        $pathToSave
+    )
+    {
         $this->countryCode = $countryCode;
         $this->whishlistId = $wishlistId;
         $this->client = $client;
         $this->logger = $logger;
-    }
+        $this->pathToSave = $pathToSave;
 
-    public function execute()
-    {
-        $page = 1;
-        $lastItemsContent = null;
-        $rows = [];
-        $baseUrl = $this->getBaseUrlForCountry($this->countryCode);
-        $wishlistBaseUrl = "{$baseUrl}/registry/wishlist/{$this->whishlistId}?layout=standard";
-        
-        if (!$baseUrl) {
-            throw new \InvalidArgumentException("Country code {$this->countryCode} is not supported.");
-        }
-
-        $this->logger->log("Exporting: {$wishlistBaseUrl}");
-
-        while (true) {
-            $url = "{$wishlistBaseUrl}&page={$page}";
-            $response = $this->client->get($url);
-            $responseContent = (string) $response->getBody();
-            $crawler = new Crawler($responseContent);
-            $items = $crawler->filter('[id^=item_]');
-
-            if ($response->getStatusCode() != 200 || !$items->count()) {
-                $this->logger->log('Empty content (are you sure that you set your list as public?)');
-                break;
-            }
-
-            if ($items->text() == $lastItemsContent) {
-                $this->logger->log('Current content is repeating last content');
-                break;
-            }
-
-            $items->each(function (Crawler $item) use (&$rows, $baseUrl) {
-                $name = trim($item->filter('[id^=itemName_]')->text());
-                $price = (float) str_replace('$', '', trim($item->filter('[id^=itemPrice_]')->text()));
-
-                $url = 
-                    $item->filter('[id^=itemName_]')->attr('href') ?
-                        $baseUrl . $item->filter('[id^=itemName_]')->attr('href') : 
-                        $item->filter('[id^=itemInfo_] .a-link-normal')->attr('href');
-
-                $image = trim($item->filter('[id^=itemImage_] img')->attr('src'));
-                $rows[] = array($name, $price, $url, $image);
-            });
-
-            $this->logger->log("Parsed page {$page} - Url: {$url}");
-
-            $lastItemsContent = $items->text();
-            ++$page;
-        }
-
-        $this->logger->log("Finished");
-
-        return $rows;
     }
 
     /**
-     * @param $countryCode
-     * @return string|null
+     * fetch wishlist and write CSV
      */
-    private function getBaseUrlForCountry($countryCode)
+    public function execute()
     {
-        $countryCode = strtoupper($countryCode);
-        $baseUrlsByCountry = [
-            'US' => 'http://www.amazon.com',
-            'DE' => 'http://www.amazon.de',
-            'UK' => 'http://www.amazon.co.uk',
-        ];
+        $amazonCrawler = new AmazonCrawler($this->logger, $this->client);
+        $items = $amazonCrawler->crawl($this->whishlistId, $this->countryCode);
 
-        $baseUrl = !empty($baseUrlsByCountry[$countryCode]) ? $baseUrlsByCountry[$countryCode] : null;
+        $items = array_merge([['Name', 'Price', 'Url','Image']], $items);
 
-        return $baseUrl;
+        // Saving each item row to the file
+        $fh = fopen($this->pathToSave, 'w');
+        foreach ($items as $item) {
+            fputcsv($fh, $item);
+        }
+
+        fclose($fh);
     }
+
+
 }
