+++ src/AmazonWishlistExporter/Logger/NullLogger.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace AmazonWishlistExporter\Logger;
+
+class NullLogger implements LoggerInterface
+{
+    public function log($message)
+    {
+
+    }
+}
